refactor: view 컨트롤러 리소스별로 분리
