+++ src/main/java/roomescape/controller/AdminController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class ViewController {
+public class AdminController {
 
     @GetMapping("/admin")
     public String adminPage() {
@@ -24,15 +24,5 @@ public class ViewController {
     @GetMapping("/admin/theme")
     public String themeAdminPage() {
         return "admin/theme";
-    }
-
-    @GetMapping("/reservation")
-    public String reservationUserPage() {
-        return "reservation";
-    }
-
-    @GetMapping
-    public String mainUserPage() {
-        return "index";
     }
 }
