+++ src/main/java/roomescape/controller/UserController.java
@@ -1,0 +1,18 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserController {
+
+    @GetMapping("/reservation")
+    public String reservationUserPage() {
+        return "reservation";
+    }
+
+    @GetMapping
+    public String mainUserPage() {
+        return "index";
+    }
+}
