+++ src/test/java/roomescape/controller/ThemeApiControllerTest.java
@@ -13,7 +13,7 @@ import static org.hamcrest.Matchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ThemeControllerTest {
+public class ThemeApiControllerTest {
 
     @Test
     @DisplayName("테마 조회를 정상적으로 수행한다.")
